requirements.txt
aiofiles==24.1.0
aiogram==2.8
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==3.2.16
django-bootstrap5==22.2
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
fastapi==0.115.0
flake8==5.0.4
frozenlist==1.4.1
h11==0.14.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.1.0
oauthlib==3.2.2
pycodestyle==2.9.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==2.5.0
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
redis==5.1.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.1
starlette==0.38.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
yarl==1.13.1
